Add code generation comment to top of generated files
By convention, a comment at the top of a file which matches this
regular expression:
  '^// Code generated .* DO NOT EDIT$'
indicates that a file is generated, and so it can be treated
differently by linters, code review tools, etc. This convention was
established in
https://github.com/golang/go/issues/13560#issuecomment-276866852.

This commit adds such a comment to the top of generated files.